+++ common/src/main/java/com/railwayteam/railways/content/custom_bogeys/blocks/gauge/standard/large/LargeCreateStyle0120BogeyBlock.java
@@ -15,10 +15,9 @@ public class LargeCreateStyle0120BogeyBlock extends CRBogeyBlock {
         super(props, defaultStyle, size);
     }
 
-    //fixme fix offset
     @Override
     public Vec3 getConnectorAnchorOffset() {
-        return new Vec3(0, 7 / 32f, 140 / 32f);
+        return new Vec3(0, 7 / 32f, 172 / 32f);
     }
 
     @Override